Remove deferred statements relating to deleted indexes
When an index is deleted, the deletion happens in line. Ensure
that any deferred statements relating to the index are expunged
so that the index is not unexpectedly resurrected later.

This can happen if, within a single migration, you have (with
surrounding boilerplate elided):

  operations=[
    CreateModel(fields=[("f", CharField(unique=True))]),
    AlterField(name="f", field=CharField(unique=False)),
  ]

If, however, these operations were to happen in different
migrations, ordering would be preserved and the deferred
statements from CreateModel would run before AlterField is
evaluated, so there is no issue in that case.